requirements.txt
together~=1.5.21
python-dotenv~=1.1.1
fastapi~=0.116.1